[AVX512] generate masks for fields using lookups
This is a bit tricky, but hopefully the comments and helper script
can help.

Mask for fields is selected based on three bits
* FLAGSTAT_FSUPPLEMENTARY
* FLAGSTAT_FSECONDARY
* FLAGSTAT_FPAIRED
However, we also take into account FLAGSTAT_FQCFAIL, and prepapre
two lookups for cases when FLAGSTAT_FQCFAIL is 0 and 1. Then
lookup is done twice and we get masks for failed and passing
reads. No extra checks are needed afterwards.